Add hourly view mode and two-tier update system for battery optimization
• Add new hourly view mode with graph rendering (24-hour temperature curve)
  - Toggle between daily and hourly views by tapping current temperature
  - Navigate through hourly data in 6-hour chunks (-6h to +18h range)
  - Render temperature graph with HourlyGraphRenderer using Canvas
  - Show text-based hourly view in 1-row widgets

• Implement two-tier update architecture to separate UI updates from data fetches
  - UI-only updates every 15-60 minutes (based on temp change rate)
  - Use AlarmManager for opportunistic UI updates without guaranteed wakeups
  - Use JobScheduler (Android 8+) to piggyback on system wakeups
  - Data fetches remain battery-aware (60-480 min intervals)
  - Instant UI updates from cache on all user interactions

• Add NWS hourly forecast API support
  - Fetch hourly data from NWS /forecast/hourly endpoint
  - Parse ISO 8601 timestamps with timezone offsets
  - Store in HourlyForecastEntity with source tracking

• Improve temperature precision and interpolation
  - Change HourlyForecastEntity.temperature from Int to Float
  - Add database migration 6_7 for decimal precision
  - Update TemperatureInterpolator to filter by API source
  - Display decimal temps (72.5°) on widgets with 2+ columns

• Add battery optimization components
  - DataFreshness utility for staleness checking
  - ScreenOnReceiver for UI updates on device unlock
  - UIUpdateReceiver and UIUpdateScheduler for temp-based scheduling
  - OpportunisticUpdateJobService for Android 8+ piggyback updates

• Expand WidgetStateManager with hourly view state
  - Add ViewMode enum (DAILY/HOURLY)
  - Add hourly offset management (-6h to +18h range)
  - Add hourly navigation bounds checking

• Add comprehensive architecture documentation
  - ARCHITECTURE.md with complete system design
  - HOURLY_VIEW_PLAN.md with feature implementation plan
  - IMPLEMENTATION_PROGRESS.md tracking completion status

• Update WeatherWidgetProvider for hourly mode support
  - Route update logic based on current view mode
  - Handle navigation differently for daily vs hourly views
  - Add toggle view mode handling
  - Extend hourly forecast range (24h past to 24h future)

+++ app/src/main/java/com/weatherwidget/data/local/HourlyForecastEntity.kt
@@ -10,7 +10,7 @@ data class HourlyForecastEntity(
     val dateTime: String,          // ISO 8601 format: "2024-01-15T14:00"
     val locationLat: Double,
     val locationLon: Double,
-    val temperature: Int,          // Temperature in Fahrenheit
+    val temperature: Float,        // Temperature in Fahrenheit
     val source: String,            // "NWS" or "OPEN_METEO"
     val fetchedAt: Long            // When this data was fetched
 )
